+++ includes/db.php
@@ -1,8 +1,8 @@
 <?php
-$database = 'YOUR_DATABASE_NAME';
-$host = 'YOUR_HOST';
-$username = 'YOUR_USER';
-$password = 'YOUR_PASSWORD';
+$database = 'class869_horta';
+$host = 'localhost';
+$username = 'root';
+$password = '';
 
 try {
     $conn = new PDO("mysql:host=$host;dbname=$database;charset=utf8mb4", $username, $password, [